StyleTransform bug fixes and optimizations. (#2)
We now avoid the runtime transform whenever possible, such as when the
user specifies the style as an uninterpolated template string, or when
the user provides object literals to style properties.

NOTE: In both Torq and the previous implementation, we parsed spread
attributes out-of-order with respect to the normal left-to-right JSX
attribute precedence. In general, JSX attributes are evaluated
left-to-right, with later attributes overwriting previous ones. Note the
semantics of the following:

  <div attr1={A} {...B} attr2={C} {...D} />

In React (and by convention), these attributes would be evaulated in
order from left to right, and overwritten accordingly. This behavior
is commonly relied upon when passing props as spread attributes.

While React does not combine style attributes, in Twist one would expect
the following:

  <div style={A} {...B} style={C} {...D} />

to result in a merged style of (A, B, C, D), in order. I think this
is the desired behavior, which is what this patch does. One can still
replicate the old Torq-style behavior by placing the spread attribute
first (to have later styles supercede earlier ones).

Let me know if you disagree with this rationale.

Additionally, the runtime transform has been optimized and simplified.